requirements.txt
pyspark==3.3.1
kafka-python-ng